update how types of assets._index response are handled
